requirements.txt
redis==3.5.3
requests==2.26.0
pandas==1.3.3
python-dotenv==0.19.1